+++ mrz-3-edition/chapters/3/algorithms/2/newton.m
@@ -15,20 +15,38 @@
 %}
 
 
-function [b] = baseComputation ()
-% BASECOMPUTATION Evaluation of base used by machine
+function [x, numberOfIterations, derivativeZero] = newton(f, fDerivative, startPoint, tolerance, maxIterations)
+% FIXEDPOINT: Finds a solution of f(x) = 0 with Newton method.
 %
-%  [b] = baseComputation ()
+%  [x, numberOfIterations, derivativeZero] = newton(f, fDerivative, startPoint, tolerance, maxIterations)
+%
+% Input:
+% f - 'f' function in the equation 'f(x) = 0'
+% fDerivative - derivative of f
+% startPoint - starting point of method
+% tolerance - epsilon at which stop method (i.e when |f(xn) - 0| <
+%             epsilon)
+% maxIterations - max number of iterations to execute
 %
 % Output:
-% b - base used by machine
+% x - approximation of solution of 'f(x) = 0'
+% numberOfIterations - number of iterations executed before getting
+%                      solution
+% derivativeZero - boolean to check whether function has stopped or not
+%                  because of a f'(x) = 0
 
-a = 1.0;
-b = 1.0;
-while ((a + 1.0) - a) == 0.0
-    a = 2.0 * a;
-end
-
-while (((a + b) - a) - b) ~= 0.0
-    b = b + 1.0;
+numberOfIterations = 0;
+derivativeZero = false;
+x = startPoint;
+deltaDiff = tolerance * 2;  % initialize diff
+while deltaDiff >= tolerance && numberOfIterations < maxIterations && ~derivativeZero
+    derivativeValue = feval(fDerivative, x);
+    if derivativeValue == 0
+        derivativeZero = true;
+    else
+        deltaDiff = - feval(f, x) / derivativeValue;
+        x = x + deltaDiff;
+        deltaDiff = abs(deltaDiff);
+        numberOfIterations  = numberOfIterations + 1;  % increase counter
+    end
 end